+++ twisted/dev-purge.sh
@@ -1,0 +1,3 @@
+#!/bin/bash
+
+./ve/bin/twistd --pidfile=/tmp/purge.pid -n purge -c config.yaml